discution changes

+++ src/app/admin/add-edit/add-edit.component.ts
@@ -2,7 +2,7 @@ import { Component, OnDestroy, OnInit, ViewChild } from '@angular/core';
 import { FormControl, FormGroup, Validators } from '@angular/forms';
 
 import { ActivatedRoute, Params, Router } from '@angular/router';
-import { Observable ,  Subscription ,  of } from 'rxjs';
+import { Observable, Subscription, of } from 'rxjs';
 
 import { MessageService } from '../../messages/message.service';
 import { FileUploadService } from '../../products/shared/file-upload.service';
@@ -10,6 +10,8 @@ import { ProductService } from '../../products/shared/product.service';
 import { ProductsCacheService } from '../../products/shared/products-cache.service';
 
 import { Product } from '../../models/product.model';
+import { PoupComponentComponent } from '../../shared/poup-component/poup-component.component';
+import {NgbModal, NgbActiveModal} from "@ng-bootstrap/ng-bootstrap";
 
 // we send and receive categories as {key:true},
 // but for the input field we need
@@ -39,8 +41,9 @@ export class AddEditComponent implements OnInit, OnDestroy {
     private productService: ProductService,
     public fileUploadService: FileUploadService,
     private productsCacheService: ProductsCacheService,
-    private log: MessageService
-  ) {}
+    private log: MessageService,
+    private modalService: NgbModal
+  ) { }
 
   ngOnInit(): void {
     this.setProduct();
@@ -187,12 +190,25 @@ export class AddEditComponent implements OnInit, OnDestroy {
     );
   }
 
+
+
   public onDelete() {
     if (this.mode === 'edit') {
-      this.productSubscription.unsubscribe();
-      this.productService.deleteProduct(this.product).then((res) => {
-        this.router.navigate(['/products']);
+
+      const modalRef = this.modalService.open(PoupComponentComponent);
+      modalRef.componentInstance.name = 'World';
+
+      modalRef.result.then((data) => {
+        if (data == 'Ok') {
+          this.productSubscription.unsubscribe();
+          this.productService.deleteProduct(this.product).then((res) => {
+            this.router.navigate(['/products']);
+          });
+        }
       });
+
+
+
     } else {
       this.log.addError(`Cannot delete new product`);
     }
